www/pyproject.toml
[project]
name = "www"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.61",
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
]
